requirements.txt
fastapi
uvicorn
pdfplumber
python-docx
pytesseract
pillow
sentence-transformers
faiss-cpu
transformers
torch
